requirements.txt
Django==5.0.3
djangorestframework==3.14.0
fake==0.8